+++ duckscript_sdk/src/sdk/std/collections/set_clear/mod.rs
@@ -1,0 +1,73 @@
+use crate::utils::pckg;
+use crate::utils::state::{get_handles_sub_state, mutate_set};
+use duckscript::types::command::{Command, CommandResult, Commands};
+use duckscript::types::instruction::Instruction;
+use duckscript::types::runtime::StateValue;
+use std::collections::HashMap;
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "SetClear")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["set_clear".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn requires_context(&self) -> bool {
+        true
+    }
+
+    fn run_with_context(
+        &self,
+        arguments: Vec<String>,
+        state: &mut HashMap<String, StateValue>,
+        _variables: &mut HashMap<String, String>,
+        _output_variable: Option<String>,
+        _instructions: &Vec<Instruction>,
+        _commands: &mut Commands,
+        _line: usize,
+    ) -> CommandResult {
+        if arguments.is_empty() {
+            CommandResult::Error("Set handle not provided.".to_string())
+        } else {
+            let state = get_handles_sub_state(state);
+
+            let key = arguments[0].clone();
+
+            let result = mutate_set(key, state, |set| {
+                set.clear();
+
+                Ok(None)
+            });
+
+            match result {
+                Ok(_) => CommandResult::Continue(Some("true".to_string())),
+                Err(error) => CommandResult::Error(error),
+            }
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}