+++ duckscript_sdk/src/sdk/std/var/set_by_name/mod.rs
@@ -1,0 +1,66 @@
+use crate::utils::pckg;
+use duckscript::types::command::{Command, CommandResult, Commands};
+use duckscript::types::instruction::Instruction;
+use duckscript::types::runtime::StateValue;
+use std::collections::HashMap;
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "SetByName")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["set_by_name".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn requires_context(&self) -> bool {
+        true
+    }
+
+    fn run_with_context(
+        &self,
+        arguments: Vec<String>,
+        _state: &mut HashMap<String, StateValue>,
+        variables: &mut HashMap<String, String>,
+        _output_variable: Option<String>,
+        _instructions: &Vec<Instruction>,
+        _commands: &mut Commands,
+        _line: usize,
+    ) -> CommandResult {
+        if arguments.is_empty() {
+            CommandResult::Error("Missing variable name.".to_string())
+        } else {
+            let output = if arguments.len() > 1 {
+                variables.insert(arguments[0].clone(), arguments[1].clone());
+                Some(arguments[1].clone())
+            } else {
+                variables.remove(&arguments[0]);
+                None
+            };
+
+            CommandResult::Continue(output)
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}